fix formatting and some lingering bugs

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: tmckinno <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/06/05 14:32:16 by tmckinno          #+#    #+#             */
-/*   Updated: 2017/06/05 22:29:34 by tmckinno         ###   ########.fr       */
+/*   Updated: 2017/06/09 18:55:54 by tmckinno         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
 	int	len;
